add --noremove flag
since we don't know if the output directory is dedicated to scalu
anymore, this flag allows the user to disable the default behavior so
no items of the directory are deleted

+++ scalu/src/cli/output_handling.py
@@ -1,6 +1,6 @@
 from pathlib import Path
 
-def handle(output_files, output_directory):
+def handle(output_files, output_directory, remove):
     # use scalu_out in current directory by default
     if not output_directory:
         output_directory = 'scalu_out/'
@@ -10,11 +10,11 @@ def handle(output_files, output_directory):
     if not output_directory.exists():
         output_directory.mkdir()
 
-    for child in output_directory.iterdir():
-        child.unlink()
+    if remove:
+        for child in output_directory.iterdir():
+            child.unlink()
 
     for cfg in output_files.files:
         output = output_directory.joinpath(cfg.name + '.' + cfg.type)
         output.touch()
         output.write_text(cfg.content)
-
